requirements.txt
distlib==0.3.5
filelock==3.7.1
pbr==5.9.0
Pillow==9.2.0
platformdirs==2.5.2
stevedore==4.0.0
virtualenv==20.16.2
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
